refactor(election-map): seperate selectors and related business logic
