+++ packages/election-map/components/mobile/MobileDashboardNew.js
@@ -1,17 +1,14 @@
-import { useState, useEffect, useMemo } from 'react'
-import { electionNamePairs } from '../../utils/election'
+import { useState } from 'react'
 import styled from 'styled-components'
-import Selector from './Selector'
-import ElectionSelector from './ElectionSelector'
-import ReferendumSelector from './ReferendumSelector'
+
 import ElectionVoteComparisonPanel from '../ElectionVoteComparisonPanel'
 
 import YearComparisonMenuBar from './YearComparisonMenuBar'
-import { useDistrictMapping } from '../../hook/useDistrictMapping'
 import InfoboxContainer from './InfoboxContainer'
 import { useAppSelector } from '../../hook/useRedux'
 import { useAppDispatch } from '../../hook/useRedux'
 import { electionActions } from '../../store/election-slice'
+import SelectorContainer from './SeletorsContainer'
 /**
  * @typedef {Object} NationData
  * @property {string} name
@@ -87,21 +84,6 @@ const Wrapper = styled.div`
 const ContentWrapper = styled.div`
   padding: 0 24px;
 `
-const DistrictSelectorWrapper = styled.div`
-  display: flex;
-  margin: 8px auto 4px;
-  width: 100%;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: center;
-  justify-content: left;
-  gap: 12px;
-`
-const ElectionSelectorWrapper = styled(DistrictSelectorWrapper)`
-  justify-content: left;
-  gap: 12px;
-  margin: 40px auto 0;
-`
 
 /**
  * Dashboard for new election map, created in 2023.11.20
@@ -110,279 +92,27 @@ const ElectionSelectorWrapper = styled(DistrictSelectorWrapper)`
  * 2. Rewrite state to make it more clear and clean, remove unneeded useEffect if need.
  */
 export const MobileDashboardNew = ({ onEvcSelected }) => {
-  const { districtMapping, hasDistrictMapping } = useDistrictMapping()
-
   const dispatch = useAppDispatch()
-  const { stopCompare, changeLevelControl, changeYear, resetLevelControl } =
-    electionActions
+  const { stopCompare, changeYear } = electionActions
 
   const [shouldOpenYearComparisonMenuBar, setShouldOpenYearComparisonMenuBar] =
     useState(false)
 
-  const [currentDistrictType, setCurrentDistrictType] = useState('nation')
-  const [currentCountyCode, setCurrentCountyCode] = useState('')
-  const [currentTownCode, setCurrentTownCode] = useState('')
-  const [currentConstituencyCode, setCurrentConstituencyCode] = useState('')
-  const [currentVillageCode, setCurrentVillageCode] = useState('')
-  const [currentConstituencyVillageCode, setCurrentConstituencyVillageCode] =
-    useState('')
   const electionsType = useAppSelector(
     (state) => state.election.config.electionType
   )
-  const electionSubTypes = useAppSelector(
-    (state) => state.election.config.subtypes
-  )
-  const currentElectionSubType = useAppSelector(
-    (state) => state.election.control.subtype
-  )
+
   const year = useAppSelector((state) => state.election.control.year)
   const years = useAppSelector((state) => state.election.config.years)
 
   const compareInfo = useAppSelector((state) => state.election.compare.info)
   const { compareMode } = compareInfo
-  const isConstituency =
-    electionsType === 'legislator' && currentElectionSubType.key === 'normal'
-  const allCounty = districtMapping.sub
-
-  const allTown = getAllTown(currentCountyCode)
-
-  const allVillage = getAllVillage(
-    isConstituency ? currentConstituencyCode : currentTownCode
-  )
-
-  function getAllTown(code) {
-    if (currentDistrictType === 'nation') {
-      return []
-    }
-    if (!code) {
-      return []
-    }
-
-    return allCounty?.find((item) => item.code === code)?.sub ?? []
-  }
-
-  function getAllVillage(code) {
-    if (
-      !code ||
-      currentDistrictType === 'nation' ||
-      currentDistrictType === 'county'
-    ) {
-      return []
-    }
-
-    return allTown?.find((item) => item.code === code)?.sub ?? []
-  }
-
-  /**
-   *
-   * @param {DistrictType} type
-   * @param {string} code
-   */
-  const handleOnClick = (
-    type = districtMapping.type,
-    code = districtMapping.code
-  ) => {
-    setCurrentDistrictType(type)
-    switch (type) {
-      case 'nation':
-        break
-      case 'county':
-        setCurrentCountyCode(code)
-        break
-      case 'town':
-        setCurrentTownCode(code)
-        break
-      case 'village':
-        if (isConstituency) {
-          setCurrentConstituencyVillageCode(code)
-        } else {
-          setCurrentVillageCode(code)
-        }
-        break
-      case 'constituency':
-        setCurrentConstituencyCode(code)
-        break
-      default:
-        break
-    }
-  }
 
   const handleCloseCompareMode = () => {
     setShouldOpenYearComparisonMenuBar(false)
     dispatch(stopCompare())
   }
 
-  const optionsForFirstDistrictSelector = useMemo(() => {
-    switch (electionsType) {
-      case 'mayor':
-      case 'councilMember':
-        return [...districtMapping.sub]
-      case 'referendum':
-      case 'president':
-      case 'legislator':
-        return [
-          {
-            type: districtMapping.type,
-            code: districtMapping.code,
-            name: districtMapping.name,
-          },
-          ...districtMapping.sub,
-        ]
-      default:
-        break
-    }
-  }, [electionsType, districtMapping])
-  const optionsForSecondDistrictSelector = useMemo(() => {
-    if (currentCountyCode) {
-      return [
-        { type: 'county', code: currentCountyCode, name: '-' },
-        ...allTown,
-      ]
-    }
-    return [...allTown]
-  }, [allTown, currentCountyCode])
-  const optionsForThirdDistrictSelector = useMemo(() => {
-    if (isConstituency && currentConstituencyCode) {
-      return [
-        { type: 'constituency', code: currentConstituencyCode, name: '-' },
-        ...allVillage,
-      ]
-    } else if (currentTownCode) {
-      return [{ type: 'town', code: currentTownCode, name: '-' }, ...allVillage]
-    }
-    return [...allVillage]
-  }, [allVillage, currentTownCode, currentConstituencyCode, isConstituency])
-
-  // //clean up
-  useEffect(() => {
-    if (currentDistrictType === 'nation') {
-      setCurrentCountyCode('')
-      setCurrentTownCode('')
-      setCurrentVillageCode('')
-    } else if (currentDistrictType === 'county') {
-      setCurrentTownCode('')
-      setCurrentConstituencyCode('')
-      setCurrentVillageCode('')
-    } else if (currentDistrictType === 'town') {
-      setCurrentConstituencyVillageCode('')
-      setCurrentVillageCode('')
-    } else if (currentDistrictType === 'constituency') {
-      setCurrentVillageCode('')
-      setCurrentConstituencyVillageCode('')
-    }
-  }, [currentDistrictType])
-  useEffect(() => {
-    // 為什麼需要將不相關的state `year`與 `electionType` 加入dependency?
-    // 因為會希望當年份或選制改變時，也能夠觸發dispatch `changeLevelControl`，避免infobox無法出現。
-    // 這個workaround違反了useEffect對dependency的設計原則，日後有時間需要調整。
-    let level = 0
-    switch (currentDistrictType) {
-      case 'nation':
-        dispatch(resetLevelControl())
-        return
-      case 'county':
-        level = 1
-        dispatch(
-          changeLevelControl({
-            level,
-            countyCode: currentCountyCode,
-            townCode: '',
-            villageCode: '',
-            constituencyCode: '',
-            activeCode: currentCountyCode,
-          })
-        )
-        break
-      case 'town':
-        level = 2
-        dispatch(
-          changeLevelControl({
-            level,
-            countyCode: currentCountyCode,
-            townCode: currentTownCode,
-            villageCode: '',
-            constituencyCode: '',
-            activeCode: currentTownCode,
-          })
-        )
-        break
-      case 'constituency':
-        level = 2
-        dispatch(
-          changeLevelControl({
-            level,
-            countyCode: currentCountyCode,
-            townCode: '',
-            villageCode: '',
-            constituencyCode: currentConstituencyCode,
-            activeCode: currentConstituencyCode,
-          })
-        )
-        break
-      case 'village':
-        level = 3
-        dispatch(
-          changeLevelControl({
-            level,
-            countyCode: currentCountyCode,
-            townCode: currentTownCode,
-            villageCode: isConstituency
-              ? currentConstituencyVillageCode
-              : currentVillageCode,
-            constituencyCode: currentConstituencyCode,
-            activeCode: currentVillageCode,
-          })
-        )
-        break
-
-      default:
-        break
-    }
-  }, [
-    dispatch,
-
-    year,
-    electionsType,
-    resetLevelControl,
-    currentDistrictType,
-    changeLevelControl,
-    currentCountyCode,
-    currentTownCode,
-    currentVillageCode,
-    currentConstituencyCode,
-    currentConstituencyVillageCode,
-    isConstituency,
-  ])
-  useEffect(() => {
-    if (!hasDistrictMapping) {
-      return
-    }
-
-    switch (electionsType) {
-      case 'mayor':
-        if (!currentCountyCode) {
-          setCurrentDistrictType('county')
-          setCurrentCountyCode(allCounty?.[0]?.code)
-        }
-        break
-      case 'councilMember':
-        if (!currentCountyCode) {
-          setCurrentDistrictType('county')
-          setCurrentCountyCode(allCounty?.[0]?.code)
-        }
-
-        break
-      //TODO: 中央選舉
-      case 'president':
-      case 'legislator':
-        break
-      //todo: 公投
-      case 'referendum':
-        break
-      default:
-        break
-    }
-  }, [hasDistrictMapping, electionsType, allCounty, allTown, currentCountyCode])
   const topButtonJsx = compareMode ? (
     <TopButton onClick={handleCloseCompareMode}>離開</TopButton>
   ) : (
@@ -405,21 +135,6 @@ export const MobileDashboardNew = ({ onEvcSelected }) => {
     </>
   )
 
-  const test = () => {
-    dispatch(
-      changeLevelControl({
-        level: 3,
-        countyCode: '63000',
-        townCode: '63000020',
-        villageCode: '63000020005',
-        constituencyCode: '',
-        activeCode: '63000020005',
-      })
-    )
-  }
-  if (!hasDistrictMapping) {
-    return <Wrapper>loading....</Wrapper>
-  }
   return (
     <>
       {shouldOpenYearComparisonMenuBar && (
@@ -432,59 +147,10 @@ export const MobileDashboardNew = ({ onEvcSelected }) => {
       <Wrapper isCompareMode={compareMode}>
         <TopButtonsWrapper>{topButtonJsx}</TopButtonsWrapper>
         <ContentWrapper>
-          <ElectionSelectorWrapper>
-            <ElectionSelector
-              options={electionNamePairs}
-              selectorType="electionType"
-            />
-            {electionSubTypes && (
-              <ElectionSelector
-                selectorType="electionSubType"
-                options={electionSubTypes}
-              />
-            )}
-            {electionsType === 'referendum' && !compareMode ? (
-              <ReferendumSelector></ReferendumSelector>
-            ) : null}
-          </ElectionSelectorWrapper>
-          <DistrictSelectorWrapper>
-            <Selector
-              options={optionsForFirstDistrictSelector}
-              districtCode={currentCountyCode}
-              onSelected={handleOnClick}
-              placeholderValue="台灣"
-            ></Selector>
-
-            <Selector
-              options={optionsForSecondDistrictSelector}
-              districtCode={
-                isConstituency ? currentConstituencyCode : currentTownCode
-              }
-              onSelected={handleOnClick}
-              placeholderValue="-"
-            ></Selector>
-
-            <Selector
-              options={optionsForThirdDistrictSelector}
-              districtCode={
-                isConstituency
-                  ? currentConstituencyVillageCode
-                  : currentVillageCode
-              }
-              onSelected={handleOnClick}
-              placeholderValue="-"
-            ></Selector>
-          </DistrictSelectorWrapper>
+          <SelectorContainer />
           <InfoboxContainer />
         </ContentWrapper>
-        <>
-          <button onClick={test}>測試</button>
-          <div>currentDistrictType:{currentDistrictType}</div>
-          <div>currentCountyCode:{currentCountyCode}</div>
-          <div>currentConstituencyCode:{currentConstituencyCode}</div>
-          <div>currentTownCode:{currentTownCode}</div>
-          <div>currentVillageCode:{currentVillageCode}</div>
-        </>
+
         {!compareMode && (
           <ElectionVoteComparisonPanel
             onEvcSelected={onEvcSelected}
